refactor: change emergencyprovider to modalprovider to suit its real purpose
