+++ src/components/AsignarSolicitudButton.tsx
@@ -9,7 +9,7 @@ import Link from 'next/link';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { toast } from 'sonner';
 import Modal from '@/components/Modal';
-import { useModal } from '@/context/EmergencyProvider';
+import { useModal } from '@/context/ModalProvider';
 
 type AsignarSolicitudButtonProps = {
   helpRequest: HelpRequestData;
